refactor(web): R6 slice-1 — split LibraryPage god-file into presentational components
Extract 5 presentational sub-components + shared timeAgo helper from
LibraryPage.tsx (902->643) into components/library/: LibraryShelfTabs,
LibraryToolbar, SavedBookListItem, UploadBookListItem, SavedBookGridCard.
Parent keeps all hooks/effects/derived-state (verbatim); leaf JSX delegated.
Search.tsx left as-is (already decomposed).

Pure presentational extraction, zero behavior change: markup/classes/aria/
data-book-id/i18n keys byte-identical, callbacks threaded with same args,
React keys stay at call-site, library E2E selectors preserved. Adversarial QA
clean; tsc 0 errors; 595 vitest pass. Browser gate = CI e2e (library specs).
